+++ service/src/main/java/ua/zakharvalko/service/impl/UserServiceImpl.java
@@ -19,12 +19,12 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Integer id) {
+    public void deleteUser(Long id) {
         userRepository.deleteById(id);
     }
 
     @Override
-    public User getById(Integer id) {
+    public User getById(Long id) {
         return userRepository.getById(id);
     }
